Refactor the code a little.
